bug na deleção quando base não é indexável
git-svn-id: http://svn.brlight.net/svn/lightbase-neo/trunk/LBGenerator@375 29b92fdf-8c97-4584-b987-84e8d3c556fa

+++ lbgenerator/model/context.py
@@ -107,10 +107,10 @@ class CustomContextFactory(SQLAlchemyORMContext):
         if member is None:
             return None
         if is_reg:
-            if self.index.delete(id):
-                self.session.delete(member)
-            else:
+            if self.index.is_indexable and not self.index.delete(id):
                 member = self.clear_del_data(id, member)
+            else:
+                self.session.delete(member)
         self.session.commit()
         self.session.close()
         return member
